Some optimizations in OpenMP version. Changed some integer types.

+++ omp/include/grid.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "constants.h"
+#include "stdint.h"
 
 /**
  * @brief Generates an initial grid for the simulation.
@@ -14,7 +15,7 @@
  * @param input_seed The seed value used to initialize the random number generator.
  * @return A 3D matrix representing the initial grid for the simulation. e.g. `grid[x][y][z]`
  */
-char ***gen_initial_grid(long long N, float density, int input_seed);
+char ***gen_initial_grid(uint64_t N, float density, int input_seed);
 
 /**
  * @brief Perform the simulation for the specified number of generations
@@ -25,7 +26,7 @@ char ***gen_initial_grid(long long N, float density, int input_seed);
  * @param max_generations Array of maximum generations for each species
  * @param num_generations Number of generations to simulate
  */
-void simulation(char ***grid, int N, long *max_counts, int *max_generations, int num_generations);
+void simulation(char ***grid, uint32_t N, uint64_t *max_counts, uint32_t *max_generations, uint32_t num_generations);
 
 /**
  * @brief Print the result for each species in increasing order
@@ -33,7 +34,7 @@ void simulation(char ***grid, int N, long *max_counts, int *max_generations, int
  * @param max_counts Array of maximum counts for each species
  * @param max_generations Array of maximum generations for each species
  */
-void print_result(long max_counts[N_SPECIES + 1], int max_generations[N_SPECIES + 1]);
+void print_result(uint64_t max_counts[N_SPECIES + 1], uint32_t max_generations[N_SPECIES + 1]);
 
 /**
  * @brief Print the grid
@@ -41,4 +42,4 @@ void print_result(long max_counts[N_SPECIES + 1], int max_generations[N_SPECIES 
  * @param grid 3D grid
  * @param N Size of the grid
  */
-void print_grid(char ***grid, int N);
+void print_grid(char ***grid, uint32_t N);
